StreetRun.Assets

Documentation

data Resources

Constructors

Resources

Fields

rBuildingSprites :: SpriteSheet BuildingKey Texture Seconds

rGrassSprites :: SpriteSheet GrassKey Texture Seconds

rBackStreetSprites :: Texture

rStreetSprites :: Texture

rRunnerSprites :: SpriteSheet RunnerKey Texture Seconds

rBirdSprites :: SpriteSheet BirdKey Texture Seconds

rLavaSprites :: SpriteSheet LavaKey Texture Seconds

rRockSprites :: SpriteSheet RockKey Texture Seconds

rGameMusic :: Music

rJumpSfx :: Chunk

rDuckSfx :: Chunk

rPointSfx :: Chunk

rBirdSfx :: Chunk

rHurtSfx :: Chunk

rLavaSfx :: Chunk

rQuakeSfx :: Chunk

rRockSfx :: Chunk

rDeathSfx :: Chunk

rRecoverSfx :: Chunk

rStockSfx :: Chunk

rPauseSprite :: Texture

rSpaceSprite :: Texture

rEscapeSprite :: Texture

rGameOverSprite :: Texture

rHiscoreSprite :: Texture

rTitleSprite :: Texture

rNumberSprites :: Number -> Texture

rControlsSprite :: Texture

convertSurface :: Surface -> PixelFormat -> IO Surface

Produce a new Surface based on an existing one, but optimized for blitting to the specified PixelFormat. SDL_surface contains collection of pixel used SDL_PixelFormat -- format in which pixel info is stored

loadSurface :: FilePath -> Maybe Color -> IO Surface

alphaColorDef :: Color