StreetRun.Config

Documentation

data Config

Constructors

Config Window Renderer Resources

data Resources

Constructors

Resources (SpriteSheet BuildingKey Texture Seconds) (SpriteSheet GrassKey Texture Seconds) Texture Texture (SpriteSheet RunnerKey Texture Seconds) (SpriteSheet BirdKey Texture Seconds) (SpriteSheet LavaKey Texture Seconds) (SpriteSheet RockKey Texture Seconds) Music Chunk Chunk Chunk Chunk Chunk Chunk Chunk Chunk Chunk Chunk Chunk Texture Texture Texture Texture Texture Texture (Number -> Texture) Texture